zooplatforma_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.0.6
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
django-filter==23.1
django-environ==0.9.0
drf-yasg==1.21.7
drf-spectacular==0.24.1
django-redis==5.4.0
exceptiongroup==1.2.2
flake8==7.1.1
flower==2.0.1
gunicorn==20.1.0
identify==2.6.1
idna==3.10
inflection==0.5.1
kombu==5.4.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pre-commit==3.7.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.5.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.0
watchdog==2.1.9
wcwidth==0.2.13